apps/api/pyproject.toml
[project]
name = "estate-analytics-api"
version = "0.1.0"
description = "EstateAnalytics Backend API"
authors = [{name = "Antigravity Agent", email = "agent@antigravity.google"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "orjson>=3.10.0",
    "redis>=5.2.0",
    "celery>=5.4.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "SIM"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
